% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_plot_function.R
\name{cor_plot}
\alias{cor_plot}
\title{A Corrgram Function}
\usage{
cor_plot(x)
}
\arguments{
\item{year}{in the year you want to examine}
}
\description{
This function allows you to make a corregram to examine the realtionships of continuous variables for a specific year
}
\examples{
cor_plot(1991)
}
\keyword{age}
\keyword{oldest}
