% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_function.R
\name{oldest}
\alias{oldest}
\title{An age Function}
\usage{
oldest(x)
}
\arguments{
\item{year}{in the year you want to examine}
}
\description{
This function allows you to find the oldest players playing in a given year
}
\examples{
oldest(1991)
}
\keyword{age}
\keyword{oldest}
