% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_function.R
\name{points}
\alias{points}
\title{A Points Function}
\usage{
points(x)
}
\arguments{
\item{year}{Enter in the year you want to examine}
}
\description{
This function allows you to find the players with the most points in a given year
}
\examples{
points()
}
\keyword{points}
